+++ R/install.R
@@ -85,17 +85,16 @@
 #'   documented here:
 #'   <https://www.tensorflow.org/install/pip#system-requirements>
 #'
-#' @param configure_cudnn If `install_tensorflow()` detects the platform is
-#'   Linux, an Nvidia GPU is available, and the TensorFlow version is 2.13 (the
-#'   default), it will install also install the pip package
-#'   "nvidia-cudnn-cu11==8.6.*", symlink the cudnn ".so" files so they can be
-#'   found by tensorflow, and emit instructions for how to install Nvidia
-#'   drivers to enable GPU usage.
+#' @param cuda logical `TRUE` or `FALSE`. If `install_tensorflow()` detects the platform is
+#'   Linux, an Nvidia GPU is available, and the TensorFlow version is 2.14 (the
+#'   default), it will install also install the required CUDA libraries through pip.
+#'
+#' @param metal Whether to install `tensorflow-metal` pip package on Arm Macs.
+#'   This enables tensorflow to use the GPU. Pass a string to install a specific
+#'   version like `"tensorflow-metal==0.7.*`.
 #'
 #' @param pip_ignore_installed Whether pip should ignore installed python
-#'   packages and reinstall all already installed python packages. This defaults
-#'   to `TRUE`, to ensure that TensorFlow dependencies like NumPy are compatible
-#'   with the prebuilt TensorFlow binaries.
+#'   packages and reinstall all already installed python packages.
 #'
 #' @param new_env If `TRUE`, any existing Python virtual environment and/or
 #'   conda environment specified by `envname` is deleted first.
@@ -117,19 +116,13 @@ function(method = c("auto", "virtualenv", "conda"),
          restart_session = TRUE,
          conda_python_version = NULL,
          ...,
-         configure_cudnn = NULL,
+         cuda = NULL,
+         metal = is_mac_arm64(),
          pip_ignore_installed = FALSE,
          new_env = identical(envname, "r-tensorflow"),
          python_version = NULL) {
 
   method <- match.arg(method)
-
-  if(is_mac_arm64()) {
-    if(!as.character(version) %in% c("default", "release") &&
-       !isTRUE(tryCatch(numeric_version(version) >= "2.13.0",
-                       error = function(e) NULL)))
-      stop("Only tensorflow>=2.13 supported on Arm Macs.")
-  }
 
   # verify 64-bit
   if (.Machine$sizeof.pointer != 8) {
@@ -146,27 +139,74 @@ function(method = c("auto", "virtualenv", "conda"),
            "R session that has not yet initialized Keras and TensorFlow (this is ",
            "to avoid DLL in use errors during installation)")
     }
+
+    if(grepl("gpu", as.character(version), ignore.case = TRUE))
+      warning("Caution: TensorFlow 2.10 was the last TensorFlow release that supported GPU on native-Windows. Starting with TensorFlow 2.11, you will need to install TensorFlow in WSL2, or install a CPU-only version of TensorFlow.",
+              if(identical(.Platform$GUI, "RStudio")) " For a guide on how to use RStudio with WSL2, see https://support.posit.co/hc/en-us/articles/360049776974-Using-RStudio-Server-in-Windows-WSL2")
+
+  }
+
+
+  can_use_gpu <- FALSE
+  if (is.null(cuda)) {
+
+    can_use_gpu <-
+      is_linux() &&
+      (version %in% c("default", "release") ||
+         isTRUE(extract_numeric_version(version) >= "2.14")) &&
+      tryCatch(as.logical(length(system("lspci | grep -i nvidia", intern = TRUE))),
+               warning = function(w) FALSE) # warning emitted by system for non-0 exit stat
+
+    cuda <- can_use_gpu
+
+  }
+
+  tf_package_spec <- parse_tensorflow_version(version)
+
+  if(isTRUE(cuda) && !grepl("^.*\\.whl$", tf_package_spec)) {
+    tf_package_spec <- sub("([^=<>!]*)(.*)", "\\1[and-cuda]\\2",
+                           tf_package_spec)
   }
 
   packages <- unique(c(
-    parse_tensorflow_version(version),
+    tf_package_spec,
     as.character(extra_packages)
   ))
 
-  has_gpu <- FALSE
-  if (is.null(configure_cudnn)) {
-    configure_cudnn <- FALSE
-    if (is_linux() &&
-        (version %in% c("default", "2.13") || grepl("^2\\.13", version)) &&
-        !grepl("cpu", version)) {
-      configure_cudnn <- has_gpu <- tryCatch(
-        as.logical(length(system("lspci | grep -i nvidia", intern = TRUE))),
-        warning = function(w) FALSE) # warning emitted by system for non-0 exit status
-    }
+
+  if (isTRUE(metal)) repeat {
+    tf_ver <- extract_numeric_version(tf_package_spec)
+    if(is.na(tf_ver))
+      break
+
+    if(tf_ver >= "2.14")
+      metal <- "tensorflow-metal>1.0.1"
+    else if (tf_ver >= "2.13")
+      metal <- "tensorflow-metal>=1.0.1"
+    else if (tf_ver >= "2.12")
+      metal <- "tensorflow-metal==0.8.*"
+    else
+      # https://pypi.org/project/tensorflow-metal/
+      metal <- "tensorflow-metal"
+
+    break
   }
 
-  if(configure_cudnn)
-    packages <- c(packages, "nvidia-cudnn-cu11==8.6.*")
+  python_version <- python_version %||% conda_python_version
+  if(method %in% c("auto", "virtualenv") &&
+     is.null(python_version)) {
+
+    # virtualenv_starter() picks the most recent version available, but older
+    # versions of tensorflow typically don't work with the latest Python
+    # release. In general, we're better off picking the oldest Python version available
+    # that works with the current release of tensorflow.
+    # TF 2.13 is compatible with Python <=3.11,>=3.8
+
+    available <- reticulate::virtualenv_starter(version = ">=3.9", all = TRUE)
+    # pick the smallest minor version, ignoring patchlevel
+    if(nrow(available))
+      python_version <- min(available$version[, 1:2])
+  }
 
   if (isTRUE(new_env)) {
 
@@ -182,24 +222,7 @@ function(method = c("auto", "virtualenv", "conda"),
 
   }
 
-
-  python_version <- python_version %||% conda_python_version
-  if(method %in% c("auto", "virtualenv") &&
-     is.null(python_version)) {
-
-    # virtualenv_starter() picks the most recent version available, but older
-    # versions of tensorflow typically don't work with the latest Python
-    # release. In general, we're better off picking the oldest Python version available
-    # that works with the current release of tensorflow.
-    # TF 2.13 is compatible with Python <=3.11,>=3.8
-
-    available <- reticulate::virtualenv_starter(version = ">=3.8", all = TRUE)
-    # pick the smallest minor version, ignoring patchlevel
-    if(nrow(available))
-      python_version <- min(available$version[, 1:2])
-  }
-
-  reticulate::py_install(
+  py_install_args <- list(
     packages       = packages,
     envname        = envname,
     method         = method,
@@ -210,69 +233,22 @@ function(method = c("auto", "virtualenv", "conda"),
     ...
   )
 
+  # now ignored, superseded by `cuda`
+  py_install_args$configure_cudnn <- NULL
 
-  if(!file.exists(python <- virtualenv_python(envname))) {
-    # only configure if venv, not conda
-    configure_cudnn <- FALSE
+  do.call(reticulate::py_install, py_install_args)
+
+  if(is_string(metal)) {
+    py_install_args$packages <- metal
+    tryCatch(do.call(reticulate::py_install, py_install_args),
+             error = function(e) {
+               message(e)
+               message("No suitable version of the 'tensorflow-metal' found. You can ",
+                       "use TensorFlow with CPU only, or install a previous release ",
+                       "of tensorflow that has GPU support on ARM macs with ",
+                       "`tensorflow::install_tensorflow(version = '2.13')`")
+             })
   }
-
-  if(isTRUE(configure_cudnn)) {
-
-    python <-  "/home/tomasz/.virtualenvs/r-tensorflow/bin/python"
-    # "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn"
-    cudnn_path <- get_cudnn_path(python)
-
-    # [1] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_adv_infer.so.8"
-    # [2] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_adv_train.so.8"
-    # [3] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_cnn_infer.so.8"
-    # [4] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_cnn_train.so.8"
-    # [5] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_ops_infer.so.8"
-    # [6] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn_ops_train.so.8"
-    # [7] "~/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/nvidia/cudnn/lib/libcudnn.so.8"
-    cudnn_sos <- Sys.glob(paste0(cudnn_path, "/lib/*.so*"))
-
-    # "/home/tomasz/.virtualenvs/r-tensorflow/lib/python3.8/site-packages/tensorflow/__init__.py"
-    tf_lib_path <- system2(python, c("-c", shQuote("import tensorflow as tf; print(tf.__file__)")),
-                       stderr = FALSE, stdout = TRUE)
-    tf_lib_path <- dirname(tf_lib_path)
-
-    from <- sub("^.*/site-packages/", "../", cudnn_sos)
-    to <- file.path(tf_lib_path, basename(cudnn_sos))
-    writeLines("creating symlinks:")
-    writeLines(paste("-", shQuote(to), "->", shQuote(from)))
-    file.symlink(from = from, to = to)
-
-
-    message("- To install the necessary drivers to enable GPU usage, ", appendLF = FALSE)
-    if (is_ubuntu()) {
-      is_wsl <- tryCatch(as.logical(length(system("uname -r | grep -i microsoft", intern = TRUE))),
-                         warning = function(w) FALSE)
-      keyring_url <- if(is_wsl)
-        "https://developer.download.nvidia.com/compute/cuda/repos/wsl-ubuntu/x86_64/cuda-keyring_1.0-1_all.deb"
-      else {
-        version_id <- system(". /etc/os-release && echo $VERSION_ID", intern = TRUE)
-        version_id <- sub(".", "", version_id, fixed = TRUE)
-        if(!version_id %in% c("1804", "2004", "2204")) {
-          warning("Unsupported Ubuntu version for CUDA 11.8. Only LTS version 18.04, 20.04 and 22.04 are supported")
-          version_id <- "2204"
-        }
-        sprintf("https://developer.download.nvidia.com/compute/cuda/repos/ubuntu%s/x86_64/cuda-keyring_1.0-1_all.deb", version_id)
-      }
-      instructions <- paste0(c(
-          "run the following terminal commands:",
-          paste("wget", keyring_url),
-          "sudo dpkg -i cuda-keyring_1.1-1_all.deb",
-          "sudo apt-get update",
-          "sudo apt-get -y install cuda-toolkit-11-8"
-        ), collapse = "\n" )
-    } else {
-      instructions <- c(
-          "follow the instructions for your Linux platform and install 'cuda-toolkit-11-8':\n",
-          "  https://developer.nvidia.com/cuda-11-8-0-download-archive?target_os=Linux&target_arch=x86_64")
-    }
-    message(instructions)
-  }
-
 
   cat("\nInstallation complete.\n\n")
 
@@ -284,29 +260,7 @@ function(method = c("auto", "virtualenv", "conda"),
   invisible(NULL)
 }
 
-
-get_cudnn_path <- function(python = py_discover_config()$python) {
-
-  # For TF 2.13, this assumes that someone already has cudn 11-8 installed,
-  # e.g., on ubuntu:
-  # sudo apt install cuda-toolkit-11-8
-  # also, that `python -m pip install 'nvidia-cudnn-cu11==8.6.*'`
-
-  force(python)
-  cudnn_module_path <- suppressWarnings(system2(
-    python, c("-c", shQuote("import nvidia.cudnn;print(nvidia.cudnn.__file__)")),
-    stdout = TRUE, stderr = TRUE))
-  if (!is.null(attr(cudnn_module_path, "status")) ||
-      !is_string(cudnn_module_path) ||
-      !file.exists(cudnn_module_path))
-    return()
-
-  dirname(cudnn_module_path)
-
-}
-
-
-default_version <- numeric_version("2.13")
+default_version <- numeric_version("2.14")
 
 parse_tensorflow_version <- function(version) {
   # returns unquoted string directly passable to pip, e.g 'tensorflow==2.5.*'
@@ -348,91 +302,9 @@ parse_tensorflow_version <- function(version) {
 }
 
 
-#' (Defunct) Install additional Python packages alongside TensorFlow
-#'
-#' This function is deprecated. Use the `extra_packages` argument to
-#' `install_tensorflow()` or `reticulate::py_install()` to install additional packages.
-#'
-#' @param packages Python packages to install
-#' @param conda Path to conda executable (or "auto" to find conda using the PATH
-#'   and other conventional install locations). Only used when TensorFlow is
-#'   installed within a conda environment.
-#'
-#' @export
-install_tensorflow_extras <- function(packages, conda = "auto") {
-     stop("Extra packages not installed (this function is deprecated). \n",
-          "Use the extra_packages argument to install_tensorflow() or reticulate::py_install() to ",
-          "install additional packages.")
-}
-
-install_tensorflow_mac_arm64 <- function(method = c("auto", "virtualenv", "conda"),
-                                         conda = "auto",
-                                         version = "default",
-                                         envname = NULL,
-                                         extra_packages = NULL,
-                                         restart_session = TRUE,
-                                         python_version = NULL) {
-  stopifnot(is_mac_arm64())
-  method <- match.arg(method)
-
-  if(method == "virtualenv")
-    stop("Tensorflow on Arm Macs only supported in Conda Python.",
-         " Install Miniconda with `reticulate::install_miniconda()`")
-
-  have_conda <- !is.null(tryCatch(conda_binary(conda), error = function(e) NULL))
-  if (!have_conda) {
-    stop("Tensorflow installation failed (no conda binary found)\n\n",
-         "Install Miniconda by running `reticulate::install_miniconda()`",
-         "before installing Tensorflow.\n",
-         call. = FALSE)
-  }
-
-  if(!version %in% c("default", "release"))
-    warning(
-      "Only `version = 'default'` supported on Arm Macs at this time.",
-      "Request for Tensorflow version '", version, "' ignored.")
-
-  conda <- conda_binary(conda)
-
-  install <- function(pkg, ...) {
-    message("Installing: ", paste(pkg, collapse = ", "))
-    py_install(pkg, ...,  method = "conda", conda = conda,
-               envname = envname, python_version = python_version)
-  }
-
-
-  # split extra_packages into pre and post install lists
-  # dangerous to invoke conda after pip, so everything conda goes first,
-  # with official tensorflow-* packages going last so their pinned dependancy version wins.
-  # packages like h5py o Pillow not available via pip yet for arm mac
-  # and tensorflow-hub via conda pulls in the wrong dependencies, so must come from pip.
-  extra_packages_pip <- intersect(extra_packages, c("tensorflow-hub",
-                                                    "tensorflow-datasets"))
-  extra_packages_conda <- setdiff(extra_packages, extra_packages_pip)
-
-  if(length(extra_packages_conda))
-    install(extra_packages_conda) # keras passes: c("pandas", "Pillow")
-
-  os_ver <- numeric_version(gsub("macOS [a-zA-Z ]*([0-9.]+)", "\\1",
-                                 utils::sessionInfo()$running, perl = TRUE))
-
-  if (os_ver < "12") {
-    install("tensorflow-deps==2.6", channel = c("apple", "conda-forge"))
-    install("tensorflow-macos==2.6.0", pip = TRUE)
-    install("tensorflow-metal==0.2.0", pip = TRUE)
-  } else {
-    install("tensorflow-deps", channel = c("apple", "conda-forge"))
-    install("tensorflow-macos", pip = TRUE)
-    install("tensorflow-metal", pip = TRUE)
-  }
-
-  if(length(extra_packages_pip))
-    install(extra_packages_pip, pip = TRUE)
-
-  if (restart_session &&
-      requireNamespace("rstudioapi", quietly = TRUE) &&
-      rstudioapi::hasFun("restartSession"))
-    rstudioapi::restartSession()
-
-  invisible()
+extract_numeric_version <- function(x, strict = FALSE) {
+  x <- gsub("[^0-9.]+", "", as.character(x), perl = TRUE)
+  x <- sub("^\\.+", "", x)
+  x <- sub("\\.+$", "", x)
+  numeric_version(x, strict = strict)
 }